requirements.txt
google-cloud-bigquery==3.17.2
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1
google-cloud-secret-manager==2.18.0
google-cloud-bigquery-storage==2.24.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
pyarrow==15.0.0
db-dtypes==1.2.0
fpdf2==2.7.8
plotly==5.18.0
pandas-gbq==0.22.0